Big Update! (Read Description)
Added jQuery to ease coding,
Which in turn made me want to make generators able to be dynamically
generated, so I did. It was a rather hard task, as you can see.
I’ve slightly updated the styles, and added a few more generators.
After this update I’m starting to work on exponential growth for the
costs.
(And updated the README)